+++ src-bindings/node/node.ml
@@ -138,6 +138,7 @@ module Path = struct
       val dirname : string -> string [@@js.global "@node_path.dirname"]
       val extname : string -> string [@@js.global "@node_path.extname"]
       val isAbsolute : string -> bool [@@js.global "@node_path.isAbsolute"]
+      val relative : string -> string -> string [@@js.global "@node_path.relative"]
       val join : (string list[@js.variadic]) -> string [@@js.global "@node_path.join"]]
 
   let delimiter =
@@ -173,6 +174,7 @@ module Fs = struct
       [@@js.global "@node_fs.access"]
 
       val readdir : string -> string list Promise.t [@@js.global "@node_fs.readdir"]
+      val realpath : string -> string Promise.t [@@js.global "@node_fs.realpath"]
 
       val readFile : string -> options:ReadFileOptions.t -> string Promise.t
       [@@js.global "@node_fs.readFile"]]