+++ src/ppx_tools/ppx_tools.ml
@@ -8,14 +8,16 @@ let get_preprocessed_ast path =
   | Error e -> Error e
 ;;
 
+let reparsed_code_of_ast source =
+  match Ppxlib.Ast_io.get_ast source with
+  | Impl structure ->
+    let structure = Ppxlib_ast.Selected_ast.To_ocaml.copy_structure structure in
+    Stdlib.Format.asprintf "%a" Pprintast.structure structure
+  | Intf signature ->
+    let signature = Ppxlib_ast.Selected_ast.To_ocaml.copy_signature signature in
+    Stdlib.Format.asprintf "%a" Pprintast.signature signature
+;;
+
 let get_reparsed_code_from_pp_file ~path =
-  get_preprocessed_ast path
-  |> Result.map ~f:(fun source ->
-    match Ppxlib.Ast_io.get_ast source with
-    | Impl structure ->
-      let structure = Ppxlib_ast.Selected_ast.To_ocaml.copy_structure structure in
-      Stdlib.Format.asprintf "%a" Pprintast.structure structure
-    | Intf signature ->
-      let signature = Ppxlib_ast.Selected_ast.To_ocaml.copy_signature signature in
-      Stdlib.Format.asprintf "%a" Pprintast.signature signature)
+  get_preprocessed_ast path |> Result.map ~f:reparsed_code_of_ast
 ;;